requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
loguru>=0.7.2
python-dotenv>=1.0.1
redis>=5.0.4
types-python-dateutil>=2.9.0
anyio>=4.4.0
python-dateutil>=2.9.0.post0
requests>=2.32.3
urllib3>=2.2.2
